// components/types.tsx
export interface AppTheme {
    deviceTheme: "dark" | "light";
    styling: {
        background: string;
        text: string;
        tag: string;
        textGray: string;
        textError: string;
        tagText: string;
        mainColor: string;
        mainColorGhost: string
        heartColor: string;
        heartColorGhost: string
    }
}